bet amount verification

+++ db/migrate/02_create_bets.rb
@@ -4,7 +4,7 @@ class CreateBets < ActiveRecord::Migration[4.2]
             t.integer :user_id
             t.integer :game_id
             t.string :bet_type
-            t.float :bet_amount
+            t.integer :bet_amount
             t.string :team_selected
         end
     end
